# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/issue-reports": {
      "get": {
        "operationId": "IssueReportsController_findAll",
        "parameters": [
          {
            "name": "engineerId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "IssueReports"
        ]
      },
      "post": {
        "operationId": "IssueReportsController_create",
        "parameters": [
          {
            "name": "X-Role",
            "in": "header",
            "description": "Allowed roles: ENGINEER",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateIssueReportDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "IssueReports"
        ]
      }
    },
    "/issue-reports/{id}": {
      "get": {
        "operationId": "IssueReportsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "IssueReports"
        ]
      },
      "patch": {
        "operationId": "IssueReportsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Role",
            "in": "header",
            "description": "Allowed roles: ENGINEER, OFFICER, ADMINISTRATOR",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateIssueReportDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "IssueReports"
        ]
      },
      "delete": {
        "operationId": "IssueReportsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Role",
            "in": "header",
            "description": "Allowed roles: ADMINISTRATOR, OFFICER, ENGINEER",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "IssueReports"
        ]
      }
    },
...
    "/progress-reports": {
      "get": {
        "operationId": "ProgressReportsController_findAll",
        "parameters": [
          {
            "name": "engineerId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workOrderId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "ProgressReports"
        ]
      },
      "post": {
        "operationId": "ProgressReportsController_create",
        "parameters": [
          {
            "name": "X-Role",
            "in": "header",
            "description": "Allowed roles: ENGINEER",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProgressReportDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "ProgressReports"
        ]
      }
    },
    "/progress-reports/{id}": {
      "get": {
        "operationId": "ProgressReportsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "ProgressReports"
        ]
      },
      "patch": {
        "operationId": "ProgressReportsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Role",
            "in": "header",
            "description": "Allowed roles: ENGINEER, OFFICER, ADMINISTRATOR",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProgressReportDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "ProgressReports"
        ]
      },
      "delete": {
        "operationId": "ProgressReportsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Role",
            "in": "header",
            "description": "Allowed roles: ADMINISTRATOR, OFFICER, ENGINEER",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "ProgressReports"
        ]
      }
    },
...
    "/procurement-bills": {
      "get": {
        "operationId": "ProcurementBillsController_findAll",
        "parameters": [
          {
            "name": "departmentId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "ProcurementBills"
        ]
      },
      "post": {
        "operationId": "ProcurementBillsController_create",
        "parameters": [
          {
            "name": "X-Role",
            "in": "header",
            "description": "Allowed roles: OFFICER, CFO, ADMINISTRATOR",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProcurementBillDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "ProcurementBills"
        ]
      }
    },
    "/procurement-bills/{id}": {
      "get": {
        "operationId": "ProcurementBillsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "ProcurementBills"
        ]
      },
      "patch": {
        "operationId": "ProcurementBillsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Role",
            "in": "header",
            "description": "Allowed roles: CFO, ADMINISTRATOR",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProcurementBillDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "ProcurementBills"
        ]
      },
      "delete": {
        "operationId": "ProcurementBillsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Role",
            "in": "header",
            "description": "Allowed roles: CFO, ADMINISTRATOR",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "ProcurementBills"
        ]
      }
    },
...
    "/fund-releases": {
      "get": {
        "operationId": "FundReleasesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "FundReleases"
        ]
      },
      "post": {
        "operationId": "FundReleasesController_create",
        "parameters": [
          {
            "name": "X-Role",
            "in": "header",
            "description": "Allowed roles: CFO, ADMINISTRATOR",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFundReleaseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "FundReleases"
        ]
      }
    },
    "/fund-releases/{id}": {
      "get": {
        "operationId": "FundReleasesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "FundReleases"
        ]
      },
      "patch": {
        "operationId": "FundReleasesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Role",
            "in": "header",
            "description": "Allowed roles: CFO, ADMINISTRATOR",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFundReleaseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "FundReleases"
        ]
      },
      "delete": {
        "operationId": "FundReleasesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Role",
            "in": "header",
            "description": "Allowed roles: CFO, ADMINISTRATOR",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "FundReleases"
        ]
      }
    },
...
  },
  "info": {
    "title": "API Documentation",
    "description": "InfraLynx API documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateIssueReportDto": {
        "type": "object",
        "properties": {
          "departmentId": {
            "type": "string"
          },
          "engineerId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "locationText": {
            "type": "string"
          },
          "severity": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        },
        "required": [
          "departmentId",
          "engineerId",
          "title",
          "category",
          "locationText",
          "severity",
          "status"
        ]
      },
      "UpdateIssueReportDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "severity": {
            "type": "string"
          }
        }
      },
...
      "CreateProgressReportDto": {
        "type": "object",
        "properties": {
          "departmentId": {
            "type": "string"
          },
          "engineerId": {
            "type": "string"
          },
          "workOrderId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "SUBMITTED",
              "ACKNOWLEDGED"
            ]
          }
        },
        "required": [
          "departmentId",
          "engineerId",
          "title",
          "summary",
          "status"
        ]
      },
      "UpdateProgressReportDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "SUBMITTED",
              "ACKNOWLEDGED"
            ]
          },
          "summary": {
            "type": "string"
          }
        }
      },
...
      "CreateProcurementBillDto": {
        "type": "object",
        "properties": {
          "departmentId": {
            "type": "string"
          },
          "vendor": {
            "type": "string"
          },
          "workOrderId": {
            "type": "string"
          },
          "amountLakhs": {
            "type": "number"
          },
          "gstValid": {
            "type": "boolean"
          },
          "status": {
            "type": "string",
            "enum": [
              "SUBMITTED",
              "UNDER_VERIFICATION",
              "APPROVED",
              "REJECTED",
              "PAID"
            ]
          }
        },
        "required": [
          "departmentId",
          "vendor",
          "workOrderId",
          "amountLakhs",
          "gstValid"
        ]
      },
      "UpdateProcurementBillDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "SUBMITTED",
              "UNDER_VERIFICATION",
              "APPROVED",
              "REJECTED",
              "PAID"
            ]
          },
          "gstValid": {
            "type": "boolean"
          }
        }
      },
...
      "CreateFundReleaseDto": {
        "type": "object",
        "properties": {
          "departmentId": {
            "type": "string"
          },
          "proposalId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "amountCr": {
            "type": "number"
          },
          "quarter": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "RELEASED",
              "WITHHELD",
              "REJECTED"
            ]
          },
          "notes": {
            "type": "string"
          }
        },
        "required": [
          "departmentId",
          "proposalId",
          "title",
          "amountCr",
          "quarter"
        ]
      },
      "UpdateFundReleaseDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "RELEASED",
              "WITHHELD",
              "REJECTED"
            ]
          },
          "releasedAt": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          }
        }
      },
...
    }
  }
}